Aperfeiçoamento do Projeto e Finalizações
Foram adicionada algumas funções e tratamentos de erros para um melhor funcionamento e comentários nos scripts explicando o que cada parte faz.

+++ src/model/Produto.java
@@ -2,50 +2,48 @@
 package model;
 
 public class Produto {
-    // Atributos
-    private int ID;
+    private short ID;
     private String Nome, Marca;
-    private double Valor;
-    private int Estoque;
+    private float Valor;
+    private short Estoque;
     
-    // Métodos get e set para ID
-    public int getID() {
+    public short getID() {
         return this.ID;
     }
-    public void setID(int id) {
+    
+    public void setID(short id) {
         this.ID = id;
     }
     
-    // Métodos get e set para Nome
     public String getNome() {
         return this.Nome;
     }
+    
     public void setNome(String nome) {
         this.Nome = nome;
     }
     
-    // Métodos get e set para Marca
     public String getMarca() {
         return this.Marca;
     }
+    
     public void setMarca(String marca) {
         this.Marca = marca;
     }
     
-    // Métodos get e set para Valor
-    public double getValor() {
+    public float getValor() {
         return this.Valor;
     }
-    public void setValor(double valor) {
+    
+    public void setValor(float valor) {
         this.Valor = valor;
     }
     
-    // Métodos get e set para Estoque
-    public int getEstoque() {
+    public short getEstoque() {
         return this.Estoque;
     }
-    public void setEstoque(int estoque) {
+    
+    public void setEstoque(short estoque) {
         this.Estoque = estoque;
     }
-    
 }
